requirements.txt
dj-config-url==0.1.1
dj-database-url==0.5.0
Django==3.0.5
django-appconf==1.0.4
django-crispy-forms==1.8.1
django-postgres-dropdb==2019.8.24
django-user-accounts==2.0.3
ez-setup==0.9
gunicorn==20.0.4
psycopg2==2.8.4
python-dotenv==0.10.5
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
djangorestframework==3.11.0
